Renamed channel_intro_mssages.jsx to channel_intro_messages.jsx
